pyproject.toml
[tool.poetry]
authors = [
  "Richard Reeve <richard.reeve@glasgow.ac.uk>",
  "Nathan Cummings <nathan.cummings@ukaea.uk>",
  "Dennis Reddyhoff <d.reddyhoff@sheffield.ac.uk>",
  "Kristian Zarebski <kristian.zarebski@ukaea.uk>",
  "Bruno Viola <bruno.viola@ukaea.uk>",
  "Ryan Field <ryan.field@glasgow.ac.uk>",
]
classifiers = [
  "License :: OSI Approved :: BSD License",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
  "Development Status :: 4 - Beta",
  "Topic :: Database :: Front-Ends",
  "Environment :: Console",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Operating System :: OS Independent",
]
description = "Synchronization interface for the SCRC FAIR Data Pipeline registry"
name = "fair-cli"
version = "0.7.0"

homepage = "https://www.fairdatapipeline.org/"

repository = "https://github.com/FAIRDataPipeline/FAIR-CLI"

documentation = "https://www.fairdatapipeline.org/docs/interface/fdp/"

license = "BSD-2-Clause"
packages = [
  {include = "fair"},
]

readme = "README.md"

keywords = [
  "FAIR Data Pipeline",
  "FAIR",
  "Data Management",
  "Provenance",
]

include = ["CHANGELOG.md", "README.md", "CITATION.cff"]

[tool.poetry.dependencies]
GitPython = "^3.1.18"
Jinja2 = "^3.0.1"
PyYAML = ">=5.4.1,<7.0.0"
click = "^8.0.0"
email-validator = "^1.1.3"
pre-commit = "^2.15.0"
pydantic = {extras = ["email"], version = "^1.9.0"}
python = "^3.8.0,<4.0"
requests = "^2.23.0"
rich = ">=10.2.3,<12.0.0"
semver = "^2.13.0"
simplejson = "^3.17.5"
toml = "^0.10.2"
validators = "^0.18.2"
netCDF4 = "^1.5.8"

[tool.poetry.dev-dependencies]
bandit = "^1.7.2"
black = "^22.1"
coverage = "^6.3"
deepdiff = "^5.5.0"
flake8 = "^3.9.2"
isort = "^5.10.1"
loremipsum = "^1.0.5"
mypy = "^0.931"
poetry = "^1.1.12"
pre-commit = "^2.16.0"
pycodestyle = "^2.7.0"
pydocstyle = "^6.0.0"
pylama = "^7.7.1"
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-dependency = "^0.5.1"
pytest-mock = "^3.7.0"
pytest-virtualenv = "^1.7.0"
requests-mock = "^1.9.3"
boto3 = "^1.24"
moto = "^4.0.3"
Flask = "^2.2.2"
Flask-Cors = "^3.0.10"

[tool.poetry.scripts]
fair = 'fair.cli:cli'

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.poetry.urls]
"Issue Tracker" = "https://github.com/FAIRDataPipeline/FAIR-CLI/issues"

[tool.poetry.extras]
all = ["moto"]